Refactor: move SlaRequestUtils to azkaban-common
So that it can be used in HttpRequestUtils
