+++ azkaban-web-server/src/main/java/azkaban/webapp/servlet/ScheduleServlet.java
@@ -26,6 +26,7 @@ import azkaban.scheduler.Schedule;
 import azkaban.scheduler.ScheduleManager;
 import azkaban.scheduler.ScheduleManagerException;
 import azkaban.server.HttpRequestUtils;
+import azkaban.server.SlaRequestUtils;
 import azkaban.server.session.Session;
 import azkaban.sla.SlaOption;
 import azkaban.user.Permission;
